Seperate GUI Components into Respective Classes
